fix add collection functionality
